use checkSession middleware

+++ api-server/routes/users.js
@@ -1,20 +1,26 @@
 const createError = require("http-errors");
 const express = require("express");
 const router = express.Router();
+
 const User = require("../models/user");
 const Session = require("../models/session");
+const { checkSession } = require("../middleware");
+
+// ===== ROUTES ===== //
 
 router.route("/").get(getAllUsers);
 
-router.route("/verify").get(handleVerification);
+router.route("/verify").get(checkSession, handleVerification);
 
 router.route("/login").post(handleLogin);
 
-router.route("/logout").post(handleLogout);
+router.route("/logout").post(checkSession, handleLogout);
 
 router.route("/signup").post(handleSignup);
 
-// SECURE ME LATER
+// ===== FUNCTIONS ===== //
+
+// SECURE ME
 async function getAllUsers(req, res, next) {
   try {
     const users = await User.find();
@@ -26,20 +32,8 @@ async function getAllUsers(req, res, next) {
 
 async function handleVerification(req, res, next) {
   try {
-    if (!("authorization" in req.headers)) {
-      throw createError(401, "Authorization header missing");
-    }
-    const { sid } = JSON.parse(req.headers.authorization);
-    // find the session
-    const session = await Session.findOne({ _id: sid });
-    if (!session || !session._id) {
-      throw createError(401, "User verification failed");
-    }
-    // get that session's user details
-    const user = await User.findOne({ _id: session.userId });
-    if (!user) {
-      throw createError(401, "User not found");
-    }
+    // user was set in checkSession middleware
+    const user = res.locals.user;
     // don't want to send the user's password to the client
     const { password, ...goodUser } = user._doc;
     res.status(200).json(goodUser);
@@ -73,9 +67,17 @@ async function handleLogin(req, res, next) {
 async function handleLogout(req, res, next) {
   try {
     const { userId } = req.body;
+    // user was set in checkSession middleware
+    const { _id } = res.locals.user;
+    if (String(_id) !== userId) {
+      throw createError(
+        401,
+        "Sorry not sorry. You can't sign out another user"
+      );
+    }
     // delete all this user's sessions
     await Session.deleteMany({ userId });
-    res.status(200).json({ message: "Successful logout" });
+    res.status(200).json({ message: "Successfully logged out" });
   } catch (err) {
     next(err);
   }
